+++ R/data_rename.R
@@ -38,8 +38,9 @@
 #' - Functions to reshape, pivot or rotate dataframes: [data_to_long()], [data_to_wide()], [data_rotate()]
 #' - Functions to rescale and reverse: [data_rescale()], [data_reverse()]
 #' - Functions to standardize, normalize, rank-transform: [standardize()], [normalize()], [ranktransform()], [winsorize()]
-#' - Split, cut and merge dataframes: [data_partition()], [data_cut()], [data_match()], [data_merge()]
-#' - Functions to find columns: [find_columns()]
+#' - Split, cut and merge dataframes: [data_partition()], [data_cut()], [data_merge()]
+#' - Functions to find or select columns: [find_columns()]
+#' - Functions to filter rows: [data_match()], [data_filter()]
 #'
 #' @export
 data_rename <- function(data, pattern = NULL, replacement = NULL, safe = TRUE, ...) {